Fix division bug in bandits.

+++ bandits.py
@@ -15,18 +15,18 @@ class gaussianArm():
     def __init__(self, mu, sigma):
         self.mean = mu
         self.var = sigma
-    
+
     def sample(self):
         R = random.gauss(self.mean, self.var)
         return R*(R>0)*(R<1) + 1.0*(R>1)
         return (random.gauss(self.mean, self.var))
-        
+
 class expArm():
     def __init__(self, lambdap):
         self.lambdap = lambdap
-        self.mean = 1/self.lambdap
-        self.var = 1/(self.lambdap)**2
-    
+        self.mean = 1./self.lambdap
+        self.var = 1./(self.lambdap)**2
+
     def sample(self):
         R = random.expovariate(self.lambdap)
         return R*(R<1) + 1.0*(R>1)
@@ -36,9 +36,8 @@ class betaArm():
     def __init__(self, alpha, beta):
         self.alpha = alpha
         self.beta = beta
-        self.mean = alpha/(alpha+beta)
-        self.var = (alpha*beta)/((alpha+beta)**2*(alpha+beta+1))
+        self.mean = alpha/float(alpha+beta)
+        self.var = (alpha*beta)/float((alpha+beta)**2*(alpha+beta+1))
     def sample(self):
         return random.betavariate(self.alpha, self.beta)
-        
-        
+
